backend/requirements.txt
sentence-transformers==2.7.0
faiss-cpu==1.8.0
gitpython==3.1.43
numpy==1.26.4
groq==1.1.2
python-dotenv==1.0.1
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
